PR#2 - adjustments, stored const in imageData.js
