+++ src/components/SetsMenu.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import { View, Text, Image, StyleSheet, FlatList, TouchableOpacity } from 'react-native';
 import { useNavigation } from '@react-navigation/native';
-import * as colors from './styles/colors'
+import styles from './styles/setsMenuStyles'
 
 const SetsMenu = () => {
   const [setsData, setSetsData] = useState([]);
@@ -44,32 +44,5 @@ const SetsMenu = () => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    padding: 10,
-  },
-  itemContainer: {
-    flex: 1,
-    margin: 5,
-    borderRadius: 10,
-    backgroundColor: colors.plainGrey,
-    alignItems: 'center',
-    justifyContent: 'center',
-    padding: 10,
-  },
-  image: {
-    width: 100,
-    height: 100,
-    borderRadius: 10,
-    marginBottom: 10,
-  },
-  text: {
-    fontSize: 16,
-    fontWeight: 'bold',
-    textAlign: 'center',
-  },
-});
 
 export default SetsMenu;
