+++ src/components/imageData.js
@@ -1,0 +1,44 @@
+const costTypeImages = {
+    Colorless: require('../../assets/colorless.png'),
+    Psychic: require('../../assets/psychic.png'),
+    Fire: require('../../assets/fire.png'),
+    Grass: require('../../assets/grass.png'),
+    Fairy: require('../../assets/fairy.png'),
+    Fighting: require('../../assets/fighting.png'),
+    Metal: require('../../assets/metal.png'),
+    Dragon: require('../../assets/dragon.png'),
+    Water: require('../../assets/water.png'),
+    Dark: require('../../assets/dark.png'),
+    Lightning: require('../../assets/lightning.png'),
+  };
+  
+  const typesImages = {
+    Colorless: require('../../assets/types/t_colorless.png'),
+    Psychic: require('../../assets/types/t_psychic.png'),
+    Fire: require('../../assets/types/t_fire.png'),
+    Grass: require('../../assets/types/t_grass.png'),
+    Fairy: require('../../assets/types/t_fairy.png'),
+    Fighting: require('../../assets/types/t_fighting.png'),
+    Metal: require('../../assets/types/t_metal.png'),
+    Dragon: require('../../assets/types/t_dragon.png'),
+    Water: require('../../assets/types/t_water.png'),
+    Dark: require('../../assets/types/t_dark.png'),
+    Lightning: require('../../assets/types/t_electric.png'),
+  };
+
+  const choiceImages = {
+    Colorless: require('../../assets/colorless.png'),
+    Psychic: require('../../assets/psychic.png'),
+    Fire: require('../../assets/fire.png'),
+    Grass: require('../../assets/grass.png'),
+    Fairy: require('../../assets/fairy.png'),
+    Fighting: require('../../assets/fighting.png'),
+    Metal: require('../../assets/metal.png'),
+    Dragon: require('../../assets/dragon.png'),
+    Water: require('../../assets/water.png'),
+    Dark: require('../../assets/dark.png'),
+    Lightning: require('../../assets/lightning.png'),
+  };
+  
+  export { costTypeImages, typesImages, choiceImages };
+
