+++ src/components/pokemonTypes/PlayCards.js
@@ -1,27 +1,15 @@
 import React, { useRef } from 'react';
 import { TouchableOpacity, Text, StyleSheet, View, Image } from 'react-native';
 import * as Animatable from "react-native-animatable";
+import { costTypeImages } from '../imageData';
+import styles from '../pokemonTypes/playStyles'
 
 const PlayCards = ({ type, onPress }) => {
-
-  const costTypeImages = {
-    Colorless: require('../../../assets/colorless.png'),
-    Psychic: require('../../../assets/psychic.png'),
-    Fire: require('../../../assets/fire.png'),
-    Grass: require('../../../assets/grass.png'),
-    Fairy: require('../../../assets/fairy.png'),
-    Fighting: require('../../../assets/fighting.png'),
-    Metal: require('../../../assets/metal.png'),
-    Dragon: require('../../../assets/dragon.png'),
-    Water: require('../../../assets/water.png'),
-    Dark: require('../../../assets/dark.png'),
-    Lightning: require('../../../assets/lightning.png'),
-  };
 
   const renderCostImage = (costType) => {
     const imageUrl = costTypeImages[costType];
     if (imageUrl) {
-      return <Image style={{ height: 60, width: 60, resizeMode: 'cover' }} source={imageUrl} />;
+      return <Image style={styles.renderCost} source={imageUrl} />;
     }
     return null;
   };
@@ -45,21 +33,6 @@ const PlayCards = ({ type, onPress }) => {
   );
 };
 
-const styles = StyleSheet.create({
-  button: {
-    paddingHorizontal: 20,
-    paddingVertical: 10,
-    backgroundColor: '#DDDDDD',
-    marginHorizontal: 5,
-    marginBottom: 10,
-    borderRadius: 8,
-    borderWidth: 1,
-    elevation: 5,
-    shadowColor: 'black', 
-  },
-  buttonText: {
-    fontSize: 18,
-  },
-});
+
 
 export default PlayCards;
